Added localisation support to the mod
NOTE: All existing assets will now be considered english assets. Another priority related note, if two mods add the same assets but if one adds it to english and the other adds it to another language, and the game ends up using the other language then the other language gets priority over english but if there is no asset for the other language but an english asset does exist then the english asset will be loaded.